requirements.txt
# cmd pip install -r requirements.txt
# python version  3.10.11
# --------------------------------------
torch==2.2.0 
torchvision==0.17.0 
torchaudio==2.2.0
PyYAML
matplotlib
rich
tqdm
tensorboard==2.10.0
torchsummary
# --------------------------------------
# git command line ( if you open the project first time)
# step 1: git checkout Wu / Chen (choose whos branch you want to checkout)
# step 2: git merge main (merge the main branch to your branch)

# git if you want to push your code
# step 1: git add .
# step 2: git commit -m "What you want to say"
# step 3: git push
